Allow default CURIs to be disabled
Previously there was no convenient way to disable the e.g. site,
vendor, and core repositories. But when developing with application
specific module install locations one might not want for modules
to be used outside of that location. This add a new environment
variables, RAKU_DISABLE_DEFAULT_CURIS, which can be set to anything
to prevent the default repositories from being added to the repo
chain itself. Note that it does still register those repositories
with the registry.