Fail gracefully when REPL is initialized without TTY
There is a widespread issue in "Bash for Git for Windows"
in that it is not in itself a proper TTY. The proposed solution
for docker, node, and others is to run the command again under
`winpty`.

We also want this to fail in other scenarios, such as if `repl()`
would be called inside of Comma's hand-rolled non-TTY "REPL".
In those cases where `winpty` would not be a usable solution,
the recommendation is to catch the exception and re-throw with
a different (or blank) solution.

Address R#4378 (#4378)